feat:Convert application to TypeScript
Chnaged the project language to TypeScript, with this brings some
restructuing of the project as a whole.

First, all code now lives in `src/`, including all .ts files, and
public files, such as stylesheets and views. This is to allow
the same reference path (e.g. import usersRouter from "./routes/users")
and have the relative path the same after transpiling into JavaScript

When building with TypeScript's `tsc` command the output .ts files
will be in `dist/` in the same directory structure as thier .ts
counterparts.

Many additional packages have been added as devDependencies, such as
typescript, eslint, rimraf. New scripts have been added and some
are chained together using `npm-run-all`. This allows `npm run dev`
to do the following:
- Use `nodemon` to watch for changes to typesript and css files
- Clean the `dist/` directory
- Lint the project with `eslint`
- Transpile all TypeScript code into JavaScript, into `dist/`
- Copy all remaining assets to the `dist/` folder so all assets
are contained within `dist/`
- Run the entrypoint file www

The dist folder is set in .gitignore and should not be added to git,
as all code to be worked on is in `src/`