feat:basic authn and authz middleware
Created two middleware functions that handle if a user is logged in
and if a user has the correct role for an endpoint. Authz additionally
checks if the user is logged in, and uses the user's session object
for user details

To store user information and roles, a database is used that stores
the oidc sub for that user, thier username, and a set role. This
defaults to a sqlite database file, db.sqlite. The database engine
and storage details can be defined with the DB_ENGINE and DB_STORAGE
environment variables